+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Using the following categories, list your changes in this order:
 -   If using `settings.py:REACTPY_DATABASE`, `reactpy_django.database.Router` must now be registered in `settings.py:DATABASE_ROUTERS`.
 -   By default, ReactPy will now use a backhaul thread to increase performance.
 -   Minimum Python version required is now `3.9`
+-   A thread-safe cache is no longer required.
 
 ## [3.2.1] - 2023-06-29